+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The changelog for [ApplozicSwift](https://github.com/AppLozic/ApplozicSwift). Al
 ### Enhancements
 
 - [AL-2856] Added a header view in chat bar to inject custom views from outside.
+- [AL-3044] Added ALKConversationListDelegate to get chat thread selection callback.
+	 And updated properties of `ALKChatViewModelProtocol` to public.
 
 ### Fixes